Adds locs picked statistics to confluence report
Adds the number of localized spots (locs) picked in the
"Find Similar" module to the Confluence report. Also, allows
quantile strings like "q0.25" to be passed to
`min_n_locs_per_frame` and `max_n_locs_per_frame` in `pick_similar`
to specify the minimum/maximum number of localizations per frame
based on the quantile of the distribution of localizations per pick.